+++ matching-service/.gitignore
@@ -1,3 +1,2 @@
 .venv
-app/__pycache__
-app/routers/__pycache__
+**/__pycache__
